requirements.txt
Django==5.1.7
whitenoise==6.7.0
python-frontmatter==1.1.0
Markdown==3.7
bleach==6.2.0
